Move color scale helper funcs to new file and types to appropriate type file
