+++ client/dom/index.ts
@@ -4,6 +4,7 @@ export * from './axisstyle'
 export * from './boxplot'
 export * from './checkbox'
 export * from './ColorScale/ColorScale.ts'
+export * from './ColorScale/helpers.ts'
 export * from './control.icons'
 export * from './downloadTextfile'
 export * from './GeneSetEdit/GeneSetEditUI'
